requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
jsonpath==0.82
more-itertools==8.7.0
openpyxl==3.0.3
packaging==20.9
pluggy==0.13.1
py==1.10.0
PyMySQL==0.9.3
pyparsing==2.4.7
pytest==5.4.2
pytest-rerunfailures==10.0
requests==2.22.0
six==1.15.0
urllib3==1.25.11
wcwidth==0.2.5
